Move components from `common` to `core` folder
