--- src/components/common/getAvatarUrl.js
@@ -1,20 +1,0 @@
-/*
-Return the image URL to be displayed inside the project card
-Can be either :
-* the GitHub owner avatar (by default if no `icon` property is specified)
-* A custom SVG file if project's `icon`property is specified.
-The SVG can be stored locally (inside `www/logos` folder) or in the cloud.
-*/
-
-const isUrl = input => input.startsWith('http')
-
-const formatIconUrl = input => (isUrl(input) ? input : `/logos/${input}`)
-
-const formatOwnerAvatar = (owner_id, size) =>
-  `https://avatars.githubusercontent.com/u/${owner_id}?v=3&s=${size}`
-
-export default function getProjectAvatarUrl(project, size) {
-  return project.icon
-    ? formatIconUrl(project.icon)
-    : formatOwnerAvatar(project.owner_id, size)
-}
