+++ src/components/core/index.js
@@ -1,7 +1,11 @@
 export * from './button'
+export * from './card'
 export * from './dropdown-menu'
 export * from './empty-content'
 export * from './grid'
+export * from './main-content'
 export * from './menu'
 export * from './popover'
+export * from './section'
+export * from './spinner'
 export * from './typography'
