+++ src/components/core/main-content.js
@@ -1,9 +1,7 @@
 import React from 'react'
 
-const MainContent = props => (
+export const MainContent = props => (
   <div id="main-content" className="container" {...props}>
     {props.children}
   </div>
 )
-
-export default MainContent
