+++ src/components/core/spinner.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 import './spinner.css'
 // import styled from "styled-components";
 
@@ -10,10 +10,8 @@ import './spinner.css'
  * A spinner component made of CSS code found here:
  * https://projects.lukehaas.me/css-loaders/
  */
-class Spinner extends Component {
-  render() {
-    return <div className="loader" />
-  }
+export const Spinner = () => {
+  return <div className="loader" />
 }
 
 export default Spinner
