+++ src/components/hall-of-fame/CardProjectLabels.js
