+++ src/components/project-details/MainTab/NpmCard/Dependencies.js
@@ -5,11 +5,11 @@ import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 import { useSelector } from 'react-redux'
 
-import StarTotal from '../../../common/utils/StarTotal'
-import ExpandableSection from './ExpandableSection'
-import DependencyTable from './DependencyTable'
+import { StarTotal } from '../../../core/project'
 import { ExternalLink } from '../../../core/typography'
 import { npmProjects } from '../../../../selectors'
+import ExpandableSection from './ExpandableSection'
+import DependencyTable from './DependencyTable'
 
 const DependenciesContainer = styled.div`
   overflow: hidden;
