+++ src/components/project-details/MainTab/ReadmeCard.js
@@ -2,8 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
-import Card from '../../common/Card'
-import Spinner from '../../common/Spinner'
+import { Card, Spinner } from '../../core'
 import { useFetchProjectReadMe } from '../../../api/hooks'
 import '../../../stylesheets/markdown-body.css'
