+++ src/components/project-list/ProjectTable.js
@@ -4,10 +4,7 @@ import styled, { css } from 'styled-components'
 import { Link } from 'react-router-dom'
 import numeral from 'numeral'
 
-import Avatar from '../common/ProjectAvatar'
-import StarDelta from '../common/utils/StarDelta'
-import StarTotal from '../common/utils/StarTotal'
-import DownloadCount from '../common/utils/DownloadCount'
+import { Avatar, DownloadCount, StarDelta, StarTotal } from '../core/project'
 import { getDeltaByDay } from '../../selectors/project'
 import TagLabelGroup from '../tags/TagLabelGroup'
 import { DropdownMenu, Button } from '../core'
