+++ src/components/search/SearchResults.js
@@ -6,9 +6,7 @@ import styled from 'styled-components'
 import { allProjects } from '../../selectors'
 import { getTagsById } from '../../selectors/tag-selectors'
 import { PaginationProvider } from '../core/pagination'
-import Spinner from '../common/Spinner'
-import { Button, PageTitle, EmptyContent } from '../core'
-import MainContent from '../common/MainContent'
+import { Button, PageTitle, EmptyContent, MainContent, Spinner } from '../core'
 import { ProjectPaginatedList } from './project-paginated-list'
 import TagLabelGroup from '../tags/TagLabelGroup'
 import { useSearch } from './SearchProvider'
